add "dvc push" on git commit and "dvc pull" on git revert
